Added `de/tuple_utils` module
